Initial setup
Initial setup to start off the fork and make it run on my device as intended.
Also added a dependency to allow  the new way of handling inputs with [LethalCompany InputUtils](https://thunderstore.io/c/lethal-company/p/Rune580/LethalCompany_InputUtils/).
Changed the manifest version to 1.0.1 as this is the current latest release on tunderstore.