Implemented InputUtils
Implemented InputUtils and reformed everything to support this new way of handling the keybinds. This now gives the mod functionality to change the keybinds in-game.

Also updated the manifest and the README to allow the owner of the mod to choose to revert to only implementing the new input utilities.